# openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/user-plants/{id}/waterings": {
      "post": {
        "operationId": "WateringController_record",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "내 식물(개체) id",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordWateringDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "물주기 기록. 같은 개체·같은 날 재기록은 기존 기록으로 동일 201(멱등 — 더블탭 안전). 타인/비존재 개체는 404(존재 은닉)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Watering_Output"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "errorCode",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "errorCode": {
                      "type": "string",
                      "enum": [
                        "UNAUTHENTICATED"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "errorCode",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "errorCode": {
                      "type": "string",
                      "enum": [
                        "USER_PLANT_NOT_FOUND"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Watering"
        ]
      },
      "get": {
        "operationId": "WateringController_list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "내 식물(개체) id",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "이전 페이지 마지막 wateredOn. 생략 시 첫 페이지. 삭제된 날짜여도 동작(존재 검사 없음)",
            "schema": {
              "format": "date",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "페이지 크기 (1–50, 기본 20)",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "물주기 이력 — 최신순(wateredOn DESC) keyset 페이지네이션. 타인/비존재 개체는 404(존재 은닉)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WateringListDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "errorCode",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "errorCode": {
                      "type": "string",
                      "enum": [
                        "UNAUTHENTICATED"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "errorCode",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "errorCode": {
                      "type": "string",
                      "enum": [
                        "USER_PLANT_NOT_FOUND"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Watering"
        ]
      }
    },
...
    "/users/me/blocks": {
      "get": {
        "operationId": "BlockController_list",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "이전 페이지 마지막 항목의 user.id. 생략 시 첫 페이지. 해제된 id여도 동작(존재 검사 없음)",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "페이지 크기 (1–50, 기본 20)",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "내가 차단한 유저 목록 — 해제 UI용. keyset 페이지네이션(user.id 오름차순). \"나를 차단한 유저\"는 노출하지 않는다 — 차단은 상대가 알 수 없어야 조용한 조치가 된다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockListDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "errorCode",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "errorCode": {
                      "type": "string",
                      "enum": [
                        "UNAUTHENTICATED"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Block"
        ]
      }
    },
    "/users/{userId}/block": {
      "put": {
        "operationId": "BlockController_block",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "차단/해제할 상대 유저 id",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "차단 — 이미 차단한 상대에 다시 보내도 동일 200(멱등). 자기 자신은 422, 없는 유저는 404. 효과는 양방향이다: 차단 후 서로의 글·댓글이 목록에서 사라진다(상세 직접 접근은 유지)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "errorCode",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "errorCode": {
                      "type": "string",
                      "enum": [
                        "UNAUTHENTICATED"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "errorCode",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "errorCode": {
                      "type": "string",
                      "enum": [
                        "BLOCK_TARGET_NOT_FOUND"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "errorCode",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 422
                    },
                    "errorCode": {
                      "type": "string",
                      "enum": [
                        "SELF_BLOCK"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Block"
        ]
      },
      "delete": {
        "operationId": "BlockController_unblock",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "차단/해제할 상대 유저 id",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "차단 해제 — 차단하지 않은 상대에 보내도 동일 200(멱등). 자기 자신·없는 유저도 200이다(PUT과 달리 422/404가 없다 — 목표 상태에 이미 도달해 있다)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "errorCode",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "errorCode": {
                      "type": "string",
                      "enum": [
                        "UNAUTHENTICATED"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Block"
        ]
      }
    },
...
  },
  "info": {
    "title": "pulmung API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RecordWateringDto": {
        "type": "object",
        "properties": {
          "wateredOn": {
            "type": "string",
            "format": "date",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$",
            "description": "물 준 날 — 유저 로컬 달력 날짜(클라가 계산). 같은 개체·같은 날 재기록은 멱등 201(기존 기록 반환)",
            "example": "2026-07-18"
          }
        },
        "required": [
          "wateredOn"
        ]
      },
      "Watering_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "wateredOn": {
            "type": "string",
            "format": "date",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$",
            "description": "물 준 날 (유저 로컬 달력 날짜)",
            "example": "2026-07-18"
          }
        },
        "required": [
          "id",
          "wateredOn"
        ],
        "additionalProperties": false,
        "id": "Watering_Output"
      },
      "WateringListDto_Output": {
        "type": "object",
        "properties": {
          "waterings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Watering_Output"
            },
            "description": "물주기 이력 — 최신순(wateredOn DESC)"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string",
                "format": "date",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
              },
              {
                "type": "null"
              }
            ],
            "description": "다음 페이지 cursor — null이면 마지막 페이지"
          }
        },
        "required": [
          "waterings",
          "nextCursor"
        ],
        "additionalProperties": false
      },
...
      "BlockListDto_Output": {
        "type": "object",
        "properties": {
          "blocks": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "user": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "nickname": {
                      "type": "string",
                      "example": "식집사"
                    }
                  },
                  "required": [
                    "id",
                    "nickname"
                  ],
                  "additionalProperties": false
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                  "description": "차단한 시각"
                }
              },
              "required": [
                "user",
                "createdAt"
              ],
              "additionalProperties": false
            },
            "description": "내가 차단한 유저들 — 해제 UI용. 정렬은 user.id 오름차순(차단 시각순이 아니다)"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              {
                "type": "null"
              }
            ],
            "description": "다음 페이지 cursor — null이면 마지막 페이지"
          }
        },
        "required": [
          "blocks",
          "nextCursor"
        ],
        "additionalProperties": false
      },
      "BlockDto_Output": {
        "type": "object",
        "properties": {
          "blocked": {
            "type": "boolean",
            "description": "요청 후 내 차단 상태 — PUT이면 항상 true, DELETE면 항상 false(멱등이라 이전 상태와 무관)"
          }
        },
        "required": [
          "blocked"
        ],
        "additionalProperties": false
      },
...
    }
  }
}